requirements.txt
alembic==1.7.5
asgiref==3.6.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Django==4.2.1
django-scheduler==0.10.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.1.3
fastapi==0.68.1
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
icalendar==5.0.5
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3
requests==2.30.0
rsa==4.9
schedule==1.2.0
six==1.16.0
SQLAlchemy==1.4.25
sqlparse==0.4.4
starlette==0.14.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.15.0
